huggingface/stable-diffusion/stable_diffusion_3_t2i/requirements.txt
--extra-index-url https://pypi.rbln.ai/simple/
optimum-rbln
protobuf>=4.25.5
sentencepiece>=0.2.0